% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{DPR_hsa04924}
\alias{DPR_hsa04924}
\title{DPR_hsa04924}
\format{
A data frame with 5,000 observations and 14 variables:
\describe{
  \item{X1}{X1}
  \item{X2}{X2}
  \item{X3}{X3}
  \item{X4}{X4}
  \item{X5}{X5}
  \item{X6}{X6}
  \item{X7}{X7}
  \item{X8}{X8}
  \item{X9}{X9}
  \item{X10}{X10}
  \item{X11}{X11}
  \item{X12}{X12}
  \item{X13}{X13}
  \item{y}{y}
}
}
\usage{
DPR_hsa04924
}
\description{
DPR_hsa04924
}
\details{
NeRiT and PMNT examples for its use.
}
\keyword{datasets}
\keyword{internal}
