% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeRiT.R
\name{NeRiT}
\alias{NeRiT}
\title{NeRiT}
\usage{
NeRiT(data1, data2, n1, formula)
}
\arguments{
\item{data1}{is the relationship among nodes:if one node links another, the relationship between them is 1, otherwise is 0.}

\item{data2}{includes all variables needed in formula.}

\item{n1}{is the number of nodes.}

\item{formula}{is outcome regression on nodes,edges and covariate with 'glm'.}
}
\value{
A data.frame.
}
\description{
NeRiT
}
\examples{
library(NeRiT)
library(KernSmooth)
library(akima)
n1 <- nrow(network_hsa04924)
formula = y~X1+X2+X3+X4+X5+X6+X7+X8+X9+X10+X11+X12+X13+V1_2+V2_3+V2_4+V4_5+V5_6+V3_7+V9_10+V5_13
outcome_PMI <- NeRiT(network_hsa04924,DPR_hsa04924,n1,formula)

}
