% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{network_hsa04924}
\alias{network_hsa04924}
\title{network_hsa04924}
\format{
A matrix
}
\usage{
network_hsa04924
}
\description{
network_hsa04924
}
\details{
See NeRiT and PMNT examples for its use.
}
\keyword{datasets}
\keyword{internal}
